requirements.txt
Django>=5.2
paho-mqtt
gunicorn
psycopg2
python-dotenv
dj-database-url
psycopg2-binary